+++ docs-src/app/App.svelte
@@ -1,10 +1,12 @@
 <Nav bind:component="{component}" />
 <main>
+	<Header />
 	<svelte:component this="{component}"/>
 </main>
 
 <script>
 import { Nav } from '../nav';
+import { Header } from '../header';
 
 
 let component = undefined;