+++ docs-src/nav/Nav.svelte
@@ -1,20 +1,9 @@
-<UIButton text
-	icon="meatballs"
+<UIButton text round
+	icon="sidebarLeft"
 	class="nav-toggler {navMobileShow ? 'visible' : ''}"
 	on:click="{toggleNav}"/>
 
 <aside class:mobile-show="{navMobileShow}">
-	<div class="nav-toolbar">
-		<div class="nav-toolbar-row">
-			<ThemeToggle round
-				class="dark-mode-switch"
-				title="Dark mode toggle"
-				items="{themes}"
-				value="{$PREFERS_DARK}"
-				on:change="{switchColorMode}"/>
-		</div>
-	</div>
-
 	<menu>
 		<h3>Intro</h3>
 		<NavItem name="Get Started" {active} />
@@ -66,19 +55,13 @@
 
 <svelte:window on:hashchange="{onhashchange}" />
 <script>
-import { Button as UIButton, ButtonToggle as ThemeToggle } from '../../src';
+import { Button as UIButton } from '../../src';
 import NavItem from './NavItem.svelte';
 import GetStarted from '../pages/start.svelte';
 import Changelog from '../pages/changelog.svelte';
-import { PREFERS_DARK } from '../../src/utils';
 import * as TestComponents from '../components';
 
-
 const components = { GetStarted, Changelog, ...TestComponents, };
-const themes = [
-	{ value: false, icon: 'sun' },
-	{ value: true, icon: 'moon' },
-];
 
 let active = location.hash.substr(1) || 'GetStarted';
 export let component = components[active];
@@ -90,11 +73,6 @@ function onhashchange () {
 	component = components[active];
 	if (window.Prism) requestAnimationFrame(() => window.Prism.highlightAll());
 	document.querySelector('main').scrollTop = 0;
-}
-
-function switchColorMode (e) {
-	console.log(111, e.detail, typeof e.detail);
-	document.documentElement.className = e.detail ? 'theme-dark' : 'theme-light';
 }